Convert project to use Bazel
This is a mega commit that converts the project over to be built by
Bazel. There isn't a terribly good reason to do this other than how it's
a build system that was made in the last 40 years.

This involved lots of BUILD file additions and updating header file
paths.

Things left to do still:

 - Cflags: `-O3 -Wall -fopenmp`
 - Update the README with Bazel instructions
