+++ raytracer/Color.cpp
