+++ raytracer/Image.cpp
