+++ raytracer/Image.h
