+++ raytracer/Light.cpp
