+++ raytracer/Light.h
