+++ raytracer/Ray.cpp
