+++ raytracer/RayTracer.cpp
@@ -1,22 +1,23 @@
 #include <limits>
-#include "RayTracer.h"
+
+#include "BSP.h"
 #include "Image.h"
-#include "Object.h"
-#include "Sphere.h"
-#include "Triangle.h"
 #include "Intersection.h"
 #include "Light.h"
-#include "Air.h"
-#include "ShinyColor.h"
-#include "FlatColor.h"
-#include "Checkerboard.h"
-#include "Marble.h"
-#include "Wood.h"
-#include "Glass.h"
-#include "NormalMap.h"
-#include "Turbulence.h"
-#include "CrissCross.h"
-#include "BSP.h"
+#include "RayTracer.h"
+#include "materials/Air.h"
+#include "materials/Checkerboard.h"
+#include "materials/CrissCross.h"
+#include "materials/FlatColor.h"
+#include "materials/Glass.h"
+#include "materials/Marble.h"
+#include "materials/NormalMap.h"
+#include "materials/ShinyColor.h"
+#include "materials/Turbulence.h"
+#include "materials/Wood.h"
+#include "objects/Object.h"
+#include "objects/Sphere.h"
+#include "objects/Triangle.h"
 
 using namespace std;
