+++ raytracer/materials/CrissCross.h
@@ -1,18 +1,23 @@
-#ifndef __CHECKERBOARD_H__
-#define __CHECKERBOARD_H__
+#ifndef __CRISS_CROSS_H__
+#define __CRISS_CROSS_H__
 
 #include "Material.h"
+#include "PerlinNoise.h"
+#include "raytracer/Color.h"
 
-class Checkerboard : public Material {
+class CrissCross : public Material {
 private:
+   PerlinNoise perlin;
+
    Color color1;
    Color color2;
+   Color color3;
    double scale;
    double shininess;
    double reflectivity;
 
 public:
-   Checkerboard(std::istream&);
+   CrissCross(std::istream&);
 
    virtual Color getColor(Vector);
    virtual double getShininess();
