+++ raytracer/materials/ShinyColor.cpp
@@ -1,7 +1,7 @@
 #include "ShinyColor.h"
 
-#include "Vector.h"
-#include "Color.h"
+#include "raytracer/Vector.h"
+#include "raytracer/Color.h"
 
 ShinyColor::ShinyColor(std::istream& in) {
    in >> color.r >> color.g >> color.b;
