+++ raytracer/materials/Wood.h
@@ -1,21 +1,22 @@
-#ifndef __CRISS_CROSS_H__
-#define __CRISS_CROSS_H__
+#ifndef __WOOD_H__
+#define __WOOD_H__
 
 #include "Material.h"
+#include "PerlinNoise.h"
+#include "raytracer/Color.h"
 
-class CrissCross : public Material {
+class Wood : public Material {
 private:
    PerlinNoise perlin;
 
    Color color1;
    Color color2;
-   Color color3;
    double scale;
    double shininess;
    double reflectivity;
 
 public:
-   CrissCross(std::istream&);
+   Wood(std::istream&);
 
    virtual Color getColor(Vector);
    virtual double getShininess();
