+++ raytracer/objects/Object.cpp
