+++ raytracer/objects/Triangle.h
@@ -3,11 +3,12 @@
 
 #include <math.h>
 #include <stdlib.h>
-#include "Vector.h"
-#include "Ray.h"
-#include "Intersection.h"
+
 #include "Object.h"
-#include "Boundaries.h"
+#include "raytracer/Boundaries.h"
+#include "raytracer/Intersection.h"
+#include "raytracer/Ray.h"
+#include "raytracer/Vector.h"
 
 class Material;
